refactor :: file name, emailSend param
- 파일 오타 수정 및 import, export 경로 수정
- emailSend에 넘겨주는 값 변경
